Update AsyncTcpDispatcher.cs
1. Fix:
Somehow the SocketAsyncEventArgs were null for me sometime, so I checked if it is null...

2.Fix
If trying to get the permissionlist it took from 90 to 120 Seconds (on localhost) to get it.
The problem was the regex checking every time the buffer was full.
Now I check if the statusline could be reached before trying to get it with regex, which increases the performance to a few seconds...

+++ TS3QueryLib.Core.Silverlight/AsyncTcpDispatcher.cs
@@ -135,7 +135,8 @@ namespace TS3QueryLib.Core
                 {
                     Socket.Shutdown(SocketShutdown.Both);
                     Socket = null;
-                    SocketAsyncEventArgs.Dispose();
+                    if (SocketAsyncEventArgs != null) //somehow I managed to throw an Exception on the next line...
+                        SocketAsyncEventArgs.Dispose();
                     SocketAsyncEventArgs = null;
                 }
             }
@@ -225,27 +226,32 @@ namespace TS3QueryLib.Core
                             continue;
                         }
 
-                        Match statusLineMatch = StatusLineMatch(_receiveRepository.ToString());
+                        
 
-                        if (statusLineMatch.Success)
+                        if(_receiveRepository.ToString().Contains("error id="))//performance fix (mainly for permissionlist)
                         {
-                            ModifyLastCommandResponse(statusLineMatch.Value);
-                            _receiveRepository.Remove(0, statusLineMatch.Length);
-
-                            SimpleResponse response = SimpleResponse.Parse(statusLineMatch.Value);
-
-                            if (response.IsBanned)
+                            Match statusLineMatch = StatusLineMatch(_receiveRepository.ToString());
+    
+                            if (statusLineMatch.Success)
                             {
-                                ThreadPool.QueueUserWorkItem(OnBanDetected, response);
-                                Disconnect();
-                                OnSocketError(System.Net.Sockets.SocketError.ConnectionReset);
-                                return;
+                                ModifyLastCommandResponse(statusLineMatch.Value);
+                                _receiveRepository.Remove(0, statusLineMatch.Length);
+    
+                                SimpleResponse response = SimpleResponse.Parse(statusLineMatch.Value);
+    
+                                if (response.IsBanned)
+                                {
+                                    ThreadPool.QueueUserWorkItem(OnBanDetected, response);
+                                    Disconnect();
+                                    OnSocketError(System.Net.Sockets.SocketError.ConnectionReset);
+                                    return;
+                                }
+    
+                                if (_receiveRepository.Length == 0)
+                                    break;
+    
+                                continue;
                             }
-
-                            if (_receiveRepository.Length == 0)
-                                break;
-
-                            continue;
                         }
 
                         break;
